[Refactor] Move paste handling to usePaste composable
